+++ src/exactly_lib/cli_default/program_modes/test_case/phases/assert_.py
@@ -2,7 +2,7 @@ from exactly_lib.common.instruction_setup import instruction_set_from_name_and_s
 from exactly_lib.definitions.test_case.instructions import instruction_names
 from exactly_lib.instructions.assert_ import contents_of_dir
 from exactly_lib.instructions.assert_ import define_symbol, change_dir, \
-    contents_of_file, env, run, exitcode, new_file, new_dir, \
+    contents_of_file, env, run, exitcode, new_file, new_dir, copy, \
     shell, existence_of_file, stdout, stderr, sys_cmd
 
 INSTRUCTIONS = instruction_set_from_name_and_setup_constructor_list(
@@ -17,6 +17,7 @@ INSTRUCTIONS = instruction_set_from_name_and_setup_constructor_list(
         (instruction_names.CHANGE_DIR_INSTRUCTION_NAME, change_dir.setup),
         (instruction_names.NEW_FILE_INSTRUCTION_NAME, new_file.setup),
         (instruction_names.NEW_DIR_INSTRUCTION_NAME, new_dir.setup),
+        (instruction_names.COPY_INSTRUCTION_NAME, copy.setup),
         (instruction_names.ENV_VAR_INSTRUCTION_NAME, env.setup),
         (instruction_names.RUN_INSTRUCTION_NAME, run.setup),
         (instruction_names.SYS_CMD_INSTRUCTION_NAME, sys_cmd.setup),