+++ test/exactly_lib_test/instructions/multi_phase/z_package_suite.py
@@ -3,6 +3,7 @@ import unittest
 from exactly_lib_test.instructions.multi_phase import \
     new_dir, change_dir, \
     env, shell, sys_cmd
+from exactly_lib_test.instructions.multi_phase.copy import z_package_suite as copy
 from exactly_lib_test.instructions.multi_phase.define_symbol import z_package_suite as define_symbol
 from exactly_lib_test.instructions.multi_phase.new_file import z_package_suite as new_file
 from exactly_lib_test.instructions.multi_phase.run_program import z_package_suite as run_program
@@ -15,6 +16,7 @@ def suite() -> unittest.TestSuite:
         new_dir.suite(),
         change_dir.suite(),
         new_file.suite(),
+        copy.suite(),
         run_program.suite(),
         env.suite(),
         sys_cmd.suite(),